apps/ai/ai_service/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=1.10.7
pydantic-extra-types==2.0.0
email-validator==2.0.0.post2
python-jose>=3.3.0
python-multipart>=0.0.6
sqlalchemy>=2.0.9
alembic>=1.10.3
psycopg2-binary>=2.9.6
pyyaml==6.0.1
jsonschema==4.18.6
tenacity==8.2.2
httpx>=0.24.0
openai>=0.27.0
tiktoken==0.5.2
# For text processing and analysis
langchain==0.1.0
nltk==3.8.1
spacy==3.7.2
python-docx==1.0.1
pypdf2==3.0.1
pyjwt>=2.6.0
python-dotenv>=1.0.0
asyncpg>=0.27.0 